Some folder work

+++ edit_folders.php
@@ -26,32 +26,43 @@ require_once( "include/message_functions.php" );
 
 
 function make_folder_form_row(&$form, $name, $nr,
-                              $bgred, $bggreen, $bgblue, $bgalpha, $fgred, $fggreen, $fgblue)
+                              $bgred, $bggreen, $bgblue, $bgalpha, $fgred, $fggreen, $fgblue,
+                              $onstatuspage)
 {
    $name_row = '<td bgcolor="#' . blend_alpha($bgred, $bggreen, $bgblue, $bgalpha) . '">' .
       '<font color="#' . blend_alpha($fgred, $fggreen, $fgblue, 255) . '">' .
       ( empty($name) ? T_('Folder name') : $name ) . '</font></td>';
 
-   $form->add_row( array( 'OWNHTML', $name_row,
-                          'TEXTINPUT', "folder$nr", 32, 32, "$name",
-                          'DESCRIPTION', T_('Backgrund'),
-                          'DESCRIPTION', T_('Red'),
-                          'TEXTINPUT', "bgred$nr", 3, 3, "$bgred",
-                          'DESCRIPTION', T_('Green'),
-                          'TEXTINPUT', "bggreen$nr", 3, 3, "$bggreen",
-                          'DESCRIPTION', T_('Blue'),
-                          'TEXTINPUT', "bgblue$nr", 3, 3, "$bgblue",
-                          'DESCRIPTION', T_('Alpha'),
-                          'TEXTINPUT', "bgalpha$nr", 3, 3, "$bgalpha",
-                          'OWNHTML', '<tr><td colspan=2>',
-                          'DESCRIPTION', T_('Foregrund'),
-                          'DESCRIPTION', T_('Red'),
-                          'TEXTINPUT', "fgred$nr", 3, 3, "$fgred",
-                          'DESCRIPTION', T_('Green'),
-                          'TEXTINPUT', "fggreen$nr", 3, 3, "$fggreen",
-                          'DESCRIPTION', T_('Blue'),
-                          'TEXTINPUT', "fgblue$nr", 3, 3, "$fgblue"));
-      $form->add_row( array('SPACE'));
+
+   $array = array( 'OWNHTML', $name_row,
+                   'TEXTINPUT', "folder$nr", 32, 32, "$name",
+                   'DESCRIPTION', T_('Backgrund'),
+                   'DESCRIPTION', T_('Red'),
+                   'TEXTINPUT', "bgred$nr", 3, 3, "$bgred",
+                   'DESCRIPTION', T_('Green'),
+                   'TEXTINPUT', "bggreen$nr", 3, 3, "$bggreen",
+                   'DESCRIPTION', T_('Blue'),
+                   'TEXTINPUT', "bgblue$nr", 3, 3, "$bgblue",
+                   'DESCRIPTION', T_('Alpha'),
+                   'TEXTINPUT', "bgalpha$nr", 3, 3, "$bgalpha" );
+
+   if( $nr <= 5 )
+      array_push( $array, 'OWNHTML', '<tr><td colspan=2>' );
+   else
+      array_push( $array,  'OWNHTML', '<tr><td><td>',
+                  'CHECKBOX', "onstatuspage$nr", 't',
+                  T_('Show on status page'), $onstatuspage );
+
+   array_push( $array, 'DESCRIPTION', T_('Foregrund'),
+               'DESCRIPTION', T_('Red'),
+               'TEXTINPUT', "fgred$nr", 3, 3, "$fgred",
+               'DESCRIPTION', T_('Green'),
+               'TEXTINPUT', "fggreen$nr", 3, 3, "$fggreen",
+               'DESCRIPTION', T_('Blue'),
+               'TEXTINPUT', "fgblue$nr", 3, 3, "$fgblue");
+
+   $form->add_row( $array );
+   $form->add_row( array('SPACE'));
 }
 
 {
@@ -62,11 +73,13 @@ function make_folder_form_row(&$form, $name, $nr,
    if( !$logged_in )
       error("not_logged_in");
 
+   $my_id = $player_row['ID'];
 
-   $folders = get_folders($player_row['ID']);
+   $folders = get_folders($my_id);
    $max_folder = array_reduce(array_keys($folders), "max");
-
-
+   $status_page_folders = empty($player_row['StatusFolders']) ? array() :
+      explode( ',', $player_row['StatusFolders'] );
+   $old_statusfolders = $status_page_folders;
 
    // Update folders
 
@@ -89,6 +102,15 @@ function make_folder_form_row(&$form, $name, $nr,
       $fgcolor = sprintf("%02x%02x%02x", $fgred, $fggreen, $fgblue);
       $name = $_POST["folder$nr"];
 
+      $onstatuspage = ( $_POST["onstatuspage$nr"] == 't' );
+      if( $nr > 5 and ( in_array($nr, $old_statusfolders) xor $onstatuspage ) )
+      {
+         if( $onstatuspage )
+            array_push($status_page_folders, $nr);
+         else
+            unset($status_page_folders[$nr]);
+      }
+
       if( empty($name) )
       {
          if( $nr > $max_folder )
@@ -109,12 +131,11 @@ function make_folder_form_row(&$form, $name, $nr,
 
       $delete = (($nr <= 5 and $STANDARD_FOLDERS[$nr] === $newfolder) or
                  ($nr > 5 and empty($name) and $is_old and
-                  folder_is_removable($nr, $player_row['ID'])));
+                  folder_is_removable($nr, $my_id)));
 
       if( $delete )
       {
-         $query = "DELETE FROM Folders " .
-            "WHERE uid='" . $player_row['ID'] . "' AND Folder_nr='$nr' LIMIT 1";
+         $query = "DELETE FROM Folders WHERE uid='$my_id' AND Folder_nr='$nr' LIMIT 1";
       }
       else if( $is_old )
       {
@@ -128,12 +149,12 @@ function make_folder_form_row(&$form, $name, $nr,
             continue;
 
          $query .= implode(",", $updates);
-         $query .= " WHERE uid='" . $player_row['ID'] . "' AND Folder_nr='$nr' LIMIT 1";
+         $query .= " WHERE uid='$my_id' AND Folder_nr='$nr' LIMIT 1";
       }
       else
       {
          $query = "INSERT INTO Folders SET " .
-            "uid='" . $player_row['ID'] . "', " .
+            "uid='$my_id', " .
             "Folder_nr='$nr', " .
             "Name='$name', " .
             "BGColor='$bgcolor', " .
@@ -144,7 +165,14 @@ function make_folder_form_row(&$form, $name, $nr,
    }
 
 
-   $folders = get_folders($player_row['ID']);
+   if( $status_page_folders != $old_statusfolders )
+   {
+      mysql_query("UPDATE Players SET StatusFolders='" .
+                  implode(',',$status_page_folders) . "' WHERE ID=$my_id LIMIT 1")
+         or die(mysql_error());
+   }
+
+   $folders = get_folders($my_id);
    $max_folder = array_reduce(array_keys($folders), "max");
 
    start_page(T_("Edit message folders"), true, $logged_in, $player_row );
@@ -166,7 +194,8 @@ function make_folder_form_row(&$form, $name, $nr,
       $fgblue = base_convert(substr($fgcolor, 4, 2), 16, 10);
 
       make_folder_form_row($form, $name, $nr,
-                           $bgred, $bggreen, $bgblue, $bgalpha, $fgred, $fggreen, $fgblue);
+                           $bgred, $bggreen, $bgblue, $bgalpha, $fgred, $fggreen, $fgblue,
+                           in_array($nr, $status_page_folders));
    }
 
 
@@ -176,7 +205,7 @@ function make_folder_form_row(&$form, $name, $nr,
 
    for($i=$max_folder+1; $i<=$max_folder+3; $i++)
    {
-      make_folder_form_row($form, '', $i, 247, 245, 227, 255, 0, 0, 0);
+      make_folder_form_row($form, '', $i, 247, 245, 227, 255, 0, 0, 0, false);
    }
